fix(repo): @Modifying return type long → int (Spring Data limitation)
Spring Data JPA solo acepta void/int/Integer como return type para
@Modifying queries (no long/Long). El caller del controller usa long
y la conversión implícita int→long funciona.

+++ src/main/java/com/imap/bpm/infrastructure/repository/JobExecutorRepository.java
@@ -43,5 +43,5 @@ public interface JobExecutorRepository extends JpaRepository<JobExecutor, UUID> 
     @Modifying
     @Transactional
     @Query("DELETE FROM JobExecutor j WHERE j.processinstanceId = :pid")
-    long deleteByProcessinstanceId(@Param("pid") UUID processinstanceId);
+    int deleteByProcessinstanceId(@Param("pid") UUID processinstanceId);
 }
